move data transformers into tc-ui
